requirements.txt
Django>=4.0
gunicorn>=20.0.4
torch>=1.10.0
transformers>=4.10.0
diffusers>=0.10.0
Pillow>=9.0.0
accelerate>=0.13.0